get memory usage for any process. fixes #988 (only linux).
