+++ lib/os.js
@@ -24,6 +24,7 @@ var binding = process.binding('os');
 exports.hostname = binding.getHostname;
 exports.loadavg = binding.getLoadAvg;
 exports.uptime = binding.getUptime;
+exports.memoryUsage = binding.getMemoryUsage;
 exports.freemem = binding.getFreeMem;
 exports.totalmem = binding.getTotalMem;
 exports.cpus = binding.getCPUs;
