+++ test/simple/test-memory-usage.js
@@ -21,8 +21,15 @@
 
 var common = require('../common');
 var assert = require('assert');
+var os = require('os');
 
-var r = process.memoryUsage();
+var r = os.memoryUsage();
 console.log(common.inspect(r));
 assert.equal(true, r['rss'] > 0);
 assert.equal(true, r['vsize'] > 0);
+
+var d = os.memoryUsage(1);
+console.log(common.inspect(r));
+assert.equal(true, d['rss'] > 0);
+assert.equal(true, d['vsize'] > 0);
+
